restructure ui directory layout.
