+++ src/app/app.component.html
@@ -2,6 +2,7 @@
   <div *ngIf="isLoading"  class="loading-overlay">
     <mat-spinner color="accent" class="spinner"></mat-spinner>
   </div>
+  <router-outlet></router-outlet>
 </div>
 
-<router-outlet></router-outlet>
+
