+++ src/app/association-switch/association-switch.component.css
