+++ src/app/association-switch/association-switch.component.spec.ts
