+++ src/app/forgotten-password/forgotten-password.component.html
@@ -4,7 +4,7 @@
       src="../assets/images/hoa-icon-white.png"
       alt="Logo"
       class="header__logo"
-      (click)="navToHome()"
+      (click)="navToLanding()"
       type="button"
     />
   </div>
