+++ src/app/header/header.component.css
@@ -1,0 +1,98 @@
+/* .header-logo {
+  max-height: 300px;
+  max-width: 100%;
+} */
+
+*{padding: 0; margin: 0; box-sizing: border-box;}
+body{
+  font-family: "Lato", sans-serif;
+  line-height: 1.7;
+}
+
+header {
+	background-image: linear-gradient(
+    to right bottom,
+    rgba(211, 211, 211, 0.8),
+    rgba(38, 41, 84, 0.8)
+  );
+	text-align: center;
+	width: 100% !important;
+	height: auto;
+	/* background-size: cover; */
+	background-attachment: center;
+	position: relative;
+	overflow: hidden;
+	border-radius: 0 0 95% 95% / 20%;
+  letter-spacing: 3.5rem;
+  color: #fff;
+  padding: 50px;
+}
+
+h1 {
+  font-family: inherit;
+  font-weight: 400;
+	font-size: 6rem;
+	margin-bottom: 30px;
+}
+
+.header-welcome {
+	margin-bottom: 30px;
+  display: block;
+  font-size: 2rem;
+  font-weight: 400;
+  letter-spacing: 1rem;
+  animation: moveInLeft 1s ease-out;
+}
+
+.header-user-name {
+	margin-bottom: 30px;
+  display: block;
+  font-size: 2rem;
+  font-weight: 400;
+  letter-spacing: 1rem;
+  animation: moveInRight 1s ease-out;
+}
+
+@keyframes moveInLeft {
+  0% {
+    opacity: 0;
+    /* transform: translateX(-100px); */
+    transform: translateX(-10rem);
+  }
+  80% {
+    /* transform: translateX(10px); */
+    transform: translateX(1rem);
+  }
+  100% {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+
+@keyframes moveInRight {
+  0% {
+    opacity: 0;
+    /* transform: translateX(100px); */
+    transform: translateX(10rem);
+  }
+  80% {
+    /* transform: translateX(-10px); */
+    transform: translateX(-1rem);
+  }
+  100% {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+
+@keyframes moveInBottom {
+  0% {
+    opacity: 0;
+    /* transform: translateY(30px); */
+    transform: translateY(3rem);
+  }
+  100% {
+    opacity: 1;
+    transform: translate(0);
+  }
+}
