+++ src/app/header/header.component.html
@@ -1,0 +1,10 @@
+<!-- <div class="header">
+  <img class="header-logo" src="/assets/images/hoa-header.jpg" alt="">
+</div> -->
+<header>
+  <div class="overlay">
+    <h1>HOA Director</h1>
+      <div class="header-welcome">Welcome!</div>
+      <div class="header-user-name">Greg</div>
+  </div>
+</header>
