+++ src/app/home/home.component.html
@@ -1,220 +1,14 @@
-<body>
-  <header class="header">
-    <!-- <div class="header__logo-box">
-    <img
-      src="../assets/images/hoa-icon-white.png"
-      alt="Logo"
-      class="header__logo"
-    />
-  </div> -->
-
-    <div class="header__text-box">
-      <!-- h1 heading is the most important heading on the page for SEO and Google to know what your page is about -->
-      <h1 class="heading-primary">
-        <span class="heading-primary--main">
-          HOA Director
-        </span>
-        <span class="heading-primary--sub">
-          a place all neighbors call home
-        </span>
-      </h1>
-
-      <!-- href='#' means leading to nothing and we give it two classes -->
-      <a
-        [routerLink]="['/login']"
-        class="btn btn--white btn--animated clickable"
-        >login</a
-      >
-    </div>
-  </header>
-
-  <!-- <main>
-    <section class="section-about"> -->
-      <!-- this is a utility class used to center this text 
-      and we can reuse this utility class to center other things in the future-->
-      <!-- <div class="u-center-text u-margin-bottom-big">
-        <h2 class="heading-secondary">
-          The community you've always dreamed of
-        </h2>
-      </div>
-
-      <div class="row">
-        <div class="col-1-of-2">
-          <h3 class="heading-tertiary u-margin-bottom-small">
-            euismod consectetur. Sed vestibulum arcu eu leo iac
-          </h3>
-          <p class="paragraph">
-            Lorem ipsum dolor, sit amet consectetur adipisicing elit. Iusto
-            cumque rem et consequuntur eos adipisci omnis corporis inventore
-            reprehenderit mollitia, magnam ab, aspernatur hic nam soluta
-            repellendus veniam eaque voluptate!
-          </p>
-          <h3 class="heading-tertiary u-margin-bottom-small">
-            Le scelerisque, felis a rutrum varius, nunc nulla placerat sapien,
-          </h3>
-          <p class="paragraph">
-            Lorem ipsum dolor sit, amet consectetur adipisicing elit.
-            Repellendus distinctio excepturi nostrum omnis repellat quas,
-            expedita animi iste consequatur quae quibusdam assumenda voluptatem
-            odio dolores fuga voluptates sed ad nobis.
-          </p>
-          <a href="#" class="btn-text">Learn More &rarr;</a>
-        </div>
-        <div class="col-1-of-2">
-          <div class="composition">
-            <img
-              src="../../assets/images/sparkler.jpg"
-              alt="Photo 1"
-              class="composition__photo composition__photo--p1"
-            />
-            <img
-              src="../../assets/images/mailbox.jpg"
-              alt="Photo 2"
-              class="composition__photo composition__photo--p2"
-            />
-            <img
-              src="../../assets/images/houses.jpg"
-              alt="Photo 3"
-              class="composition__photo composition__photo--p3"
-            />
-          </div>
+<div class="container-fluid" id="site-container">
+  <div class="row justify-content-end">
+    <div class="app-container">
+      <app-header class="app-header"></app-header>
+      <div class="row" id="content-row">
+        <app-sidebar class="col-1-of-4"></app-sidebar>
+        <div class="col-3-of-4">
+          <router-outlet></router-outlet>
         </div>
       </div>
-    </section>
-
-    <section class="section-features">
-      <div class="row">
-        <div class="col-1-of-4">
-          <div class="feature-box">
-            <i class="feature-box__icon fa-vote-yea"></i>
-            <h3 class="heading-tertiary u-margin-bottom-small">
-              Rules at Your Fingertips
-            </h3>
-            <p class="feature-box__text">
-              Can't find your association rules again. We've got your back. Find
-              all your association rules in one convenient place.
-            </p>
-          </div>
-        </div>
-
-        <div class="col-1-of-4">
-          <div class="feature-box">
-            <i class="feature-box__icon icon-basic-compass"></i>
-            <h3 class="heading-tertiary u-margin-bottom-small">
-              Love Your Community
-            </h3>
-            <p class="feature-box__text">
-              You love your home! Why not love the community too. We'll help you
-              build a strong community by streamlining communication among
-              members and expediting decisions.
-            </p>
-          </div>
-        </div>
-
-        <div class="col-1-of-4">
-          <div class="feature-box">
-            <i class="feature-box__icon icon-basic-map"></i>
-            <h3 class="heading-tertiary u-margin-bottom-small">
-              On-Demand Resolution
-            </h3>
-            <p class="feature-box__text">
-              Voting on disputes and association decisions made easy. No need to
-              gather to vote on issues. Raise a concern and vote online.
-            </p>
-          </div>
-        </div>
-
-        <div class="col-1-of-4">
-          <div class="feature-box">
-            <i class="feature-box__icon icon-basic-heart"></i>
-            <h3 class="heading-tertiary u-margin-bottom-small">
-              Coordinate Events
-            </h3>
-            <p class="feature-box__text">
-              Have fun with your neighbors. Post information about association
-              events and gatherings.
-            </p>
-          </div>
-        </div>
-      </div>
-    </section>
-  </main> -->
-
-  <!-- <section class="grid-test">
-    <div class="row">
-      <div class="col-1-of-2">
-        col-1-of-2
-      </div>
-
-      <div class="col-1-of-2">
-        col-1-of-2
-      </div>
+      <app-footer></app-footer>
     </div>
-
-    <div class="row">
-      <div class="col-1-of-3">
-        col-1-of-3
-      </div>
-
-      <div class="col-1-of-3">
-        col-1-of-3
-      </div>
-
-      <div class="col-1-of-3">
-        col-1-of-3
-      </div>
-    </div>
-
-    <div class="row">
-      <div class="col-1-of-3">
-        col-1-of-3
-      </div>
-
-      <div class="col-2-of-3">
-        col-2-of-3
-      </div>
-    </div>
-
-    <div class="row">
-      <div class="col-1-of-4">
-        col-1-of-4
-      </div>
-
-      <div class="col-1-of-4">
-        col-1-of-4
-      </div>
-
-      <div class="col-1-of-4">
-        col-1-of-4
-      </div>
-
-      <div class="col-1-of-4">
-        col-1-of-4
-      </div>
-    </div>
-
-    <div class="row">
-      <div class="col-1-of-4">
-        col-1-of-4
-      </div>
-
-      <div class="col-1-of-4">
-        col-1-of-4
-      </div>
-
-      <div class="col-2-of-4">
-        col-2-of-4
-      </div>
-    </div>
-
-    <div class="row">
-      <div class="col-1-of-4">
-        col-1-of-4
-      </div>
-
-      <div class="col-3-of-4">
-        col-3-of-4
-      </div>
-    </div>
-  </section> -->
-</body>
+  </div>
+</div>
