+++ src/app/home/home.component.spec.ts
@@ -1,4 +1,7 @@
+/* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+import { DebugElement } from '@angular/core';
 
 import { HomeComponent } from './home.component';
 
@@ -8,8 +11,9 @@ describe('HomeComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [HomeComponent],
-    }).compileComponents();
+      declarations: [ HomeComponent ]
+    })
+    .compileComponents();
   }));
 
   beforeEach(() => {
