+++ src/app/models/user.model.ts
@@ -1,0 +1,3 @@
+export interface UserModel {
+  firstName: string;
+}
