--- src/app/ui/header/header.component.css
@@ -1,4 +1,0 @@
-.header-logo {
-  max-height: 300px;
-  max-width: 100%;
-}
