--- src/app/ui/header/header.component.html
@@ -1,3 +1,0 @@
-<div class="header">
-  <img class="header-logo" src="/assets/images/hoa-header.jpg" alt="">
-</div>
