--- src/app/ui/header/header.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.css'],
-})
-export class HeaderComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit() {}
-}
