--- src/app/ui/layout/layout.component.css
@@ -1,12 +1,0 @@
-#site-container {
-  background-color: #e5e5ea;
-  min-height: 100vh;
-}
-
-#content-row {
-  min-height: 100%;
-}
-
-.app-container {
-  height: 100%;
-}
