--- src/app/ui/layout/layout.component.html
@@ -1,18 +1,0 @@
-<div class="container-fluid" id="site-container">
-  <div class="row justify-content-end">
-    <div class="app-container">
-      <app-header class="row"></app-header>
-      <div class="row" id="content-row">
-        <app-sidebar class="col-1-of-4"></app-sidebar>
-        <div class="col-3-of-4">
-          <router-outlet></router-outlet>
-        </div>
-      </div>
-      <app-footer></app-footer>
-    </div>
-  </div>
-</div>
-
-<!-- <app-header class="row"></app-header>
-<ng-content></ng-content>
-<app-footer></app-footer> -->
