--- src/app/ui/main/main.component.html
@@ -1,3 +1,0 @@
-<app-layout>
-  <ng-content></ng-content>
-</app-layout>
