--- src/app/ui/ui.module.spec.ts
@@ -1,13 +1,0 @@
-import { UiModule } from './ui.module';
-
-describe('UiModule', () => {
-  let uiModule: UiModule;
-
-  beforeEach(() => {
-    uiModule = new UiModule();
-  });
-
-  it('should create an instance', () => {
-    expect(uiModule).toBeTruthy();
-  });
-});
