--- src/app/ui/ui.module.ts
@@ -1,40 +1,0 @@
-import { CommonModule } from "@angular/common";
-import { FormsModule } from "@angular/forms";
-import { NgModule } from "@angular/core";
-import { RouterModule } from "@angular/router";
-import { FooterComponent } from "./footer/footer.component";
-import { HeaderComponent } from "./header/header.component";
-import { LayoutComponent } from "./layout/layout.component";
-import { SidebarComponent } from "./sidebar/sidebar.component";
-import { AssociationSwitchComponent } from "./association-switch/association-switch.component";
-import { MainComponent } from "./main/main.component";
-
-import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-
-import { MatButtonModule } from "@angular/material/button";
-import { MatTableModule } from "@angular/material/table";
-import { MatSelectModule } from "@angular/material/select";
-import { MatFormFieldModule } from "@angular/material/form-field";
-
-@NgModule({
-  imports: [
-    CommonModule,
-    RouterModule,
-    FormsModule,
-    BrowserAnimationsModule,
-    MatButtonModule,
-    MatTableModule,
-    MatSelectModule,
-    MatFormFieldModule
-  ],
-  declarations: [
-    LayoutComponent,
-    HeaderComponent,
-    FooterComponent,
-    SidebarComponent,
-    AssociationSwitchComponent,
-    MainComponent,
-  ],
-  exports: [LayoutComponent],
-})
-export class UiModule {}
