+++ src/styles.css
@@ -29,8 +29,8 @@ html, body { height: 100%; }
 /* body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; } */
 body { font-family: Roboto, "Helvetica Neue", sans-serif; }
 
-.row {
-  max-width: 114rem;
+.row:not(.app-header) {
+  max-width: 160rem;
   margin: 0 auto;
 }
 .row:not(:last-child) {
